requirements.txt
transformers
langchain
langchain-huggingface
pinecone
python-dotenv
langchain-openai
langchain-pinecone
